refactor: adding on hooks to register a presence component so that we can place presence provider at the top of the component tree, and added an onExitEnd prop so that parent can rerender when exit animations are complete
