+++ src/Presence/types.ts
@@ -1,6 +1,8 @@
 export interface PresenceProps {
+	id?: string;
 	initial?: boolean;
 	exitBeforeEnter?: boolean;
+	onExitEnd?: () => boolean;
 }
 
 export type ReactElementWithKey = React.ReactElement & {
