+++ src/index.tsx
@@ -1,0 +1,7 @@
+export { Motion } from "./Motion";
+export {
+	Presence,
+	PresenceProvider,
+	useRegisterPresence,
+	usePresence,
+} from "./Presence";
